interface restructuring in one file types.ts
