+++ client/src/redux/ApiSlice/publicApi.ts
@@ -1,32 +1,14 @@
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
 import {
   Address,
+  ApiResponse,
+  ApiSkillResponse,
   Education,
   Experience,
   Resources,
-  Skill,
   User,
 } from "../../types";
 const api_url: string = import.meta.env.VITE_API_HOST;
-
-export interface ApiResponse<T> {
-  success: boolean;
-  data: T;
-}
-
-export interface ApiUserResponse {
-  success: boolean;
-  data: string;
-}
-
-export interface ApiSkillResponse {
-  success: boolean;
-  data: {
-    skills: Skill[];
-    id: string;
-    username: string;
-  };
-}
 
 export const publicApi = createApi({
   reducerPath: "publicApi",
